requirements.txt
pandas
numpy
matplotlib
python-dotenv
aiohttp
httpx
loguru
streamlit
seaborn